+++ templates/show_song_previews.inc.php
@@ -22,15 +22,7 @@
 <table class="tabledata">
     <thead>
         <tr class="th-top">
-            <?php if (AmpConfig::get('echonest_api_key')) { ?>
-                <th class="cel_play"></th>
-            <?php
-} ?>
             <th class="cel_song"><?php echo T_('Song Title'); ?></th>
-            <?php if (AmpConfig::get('echonest_api_key')) { ?>
-                <th class="cel_add"></th>
-            <?php
-    } ?>
             <th class="cel_artist"><?php echo T_('Artist'); ?></th>
             <th class="cel_album"><?php echo T_('Album'); ?></th>
             <th class="cel_track"><?php echo T_('Track'); ?></th>